qc-backend-server/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements --dev
#

# Note: in pipenv 2020.x, "--dev" changed to emit both default and development
# requirements. To emit only development requirements, pass "--dev-only".

-i https://pypi.org/simple
appdirs==1.4.4
asgiref==3.4.1; python_version >= '3.6'
black==21.5b0
click==8.0.1; python_version >= '3.6'
fastapi==0.66.0
gcloud==0.17.0
googleapis-common-protos==1.53.0; python_version >= '3.6'
h11==0.12.0; python_version >= '3.6'
httplib2==0.19.1
httptools==0.2.0
jws==0.1.3
mypy-extensions==0.4.3
oauth2client==3.0.0
pathspec==0.8.1
protobuf==3.17.3
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycryptodome==3.4.3
pydantic==1.8.2; python_full_version >= '3.6.1'
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyrebase==3.0.27
python-dotenv==0.18.0
python-jwt==2.0.1
pyyaml==5.4.1
redis==3.5.3
regex==2021.7.6
requests-toolbelt==0.7.0
requests==2.11.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rsa==4.7.2; python_version >= '3.5' and python_version < '4'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
starlette==0.14.2; python_version >= '3.6'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
typing-extensions==3.10.0.0
uvicorn[standard]==0.14.0
uvloop==0.15.2
watchgod==0.7
websockets==9.1
